launch screen has bundle number automatically
